Rename beta versions to canary versions
