+++ packages/solid/project.json
@@ -42,7 +42,7 @@
         }
       ],
       "configurations": {
-        "beta": {
+        "canary": {
           "args": "--rename=canary"
         }
       }
