+++ tools/scripts/canary/canary_rename.mjs
